requirements.txt
alembic==1.1.0
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
factory-boy==2.12.0
Faker==2.0.1
filelock==3.0.12
Flask==1.1.1
Flask-Limiter==1.0.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.0
google-auth==1.6.3
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
limits==1.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.13.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.8
text-unidecode==1.2
toml==0.10.0
tox==3.14.0
urllib3==1.25.3
virtualenv==16.7.5
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
